Fixes the issue of droping the IdSplice rule.
The problem was that both the IdSplice and ParenSplice rules rewrote
Hs.Exp's. Combining them with `extQ` resulted that only one was
applied. This is fixed by a new combinator `extQA` that (in this
specific case) tries the specific case, and when it doesn't rewrite
it falls back to the general case.